+++ src/modules/Supplier/ListSupplier.jsx
@@ -28,8 +28,9 @@ import {
 import { visuallyHidden } from '@mui/utils';
 import { IconFileExport } from '@tabler/icons';
 
-import { Message, TitleButton } from 'components/helpers/Enums';
-import { SNACKBAR_OPEN } from 'store/actions';
+import swal from 'sweetalert';
+import { MessageDelete, ParamDelete } from 'components/alert/AlertAll';
+import { TitleButton } from 'components/helpers/Enums';
 import MainCard from 'ui-component/cards/MainCard';
 import { GetAllSupplier, DeleteSupplier } from 'api/clients/SupplierClient';
 
@@ -202,9 +203,9 @@ EnhancedTableToolbar.propTypes = {
 };
 
 const ListSupplier = () => {
-    const dispatch = useDispatch();
-    const [supplier, setSupplier] = useState([]);
     const navigate = useNavigate();
+    const [supplier, setSupplier] = useState([]);
+    const [openDelete, setOpenDelete] = useState(false);
     const [idCheck, setIdCheck] = useState('');
 
     const theme = useTheme();
@@ -306,20 +307,18 @@ const ListSupplier = () => {
 
     const handleDelete = async () => {
         try {
-            const result = await DeleteSupplier(idCheck);
-            if (result.status === 200) {
-                dispatch({
-                    type: SNACKBAR_OPEN,
-                    open: true,
-                    message: `${Message.Eliminar}`,
-                    variant: 'alert',
-                    alertSeverity: 'error',
-                    close: false,
-                    transition: 'SlideUp'
-                })
-            }
-            setSelected([]);
-            GetAll();
+            swal(ParamDelete).then(async (willDelete) => {
+                if (willDelete) {
+                    const result = await DeleteSupplier(idCheck);
+                    if (result.status === 200) {
+                        setOpenDelete(true);
+                    }
+                    setSearch('');
+                    setSelected([]);
+                    GetAll();
+                } else
+                    setSelected([]);
+            });
         } catch (error) {
             console.log(error);
         }
@@ -330,6 +329,7 @@ const ListSupplier = () => {
 
     return (
         <MainCard title="Lista de Proveedores" content={false}>
+            <MessageDelete open={openDelete} onClose={() => setOpenDelete(false)} />
             <CardContent>
                 <Grid container justifyContent="space-between" alignItems="center" spacing={2}>
                     <Grid item xs={12} sm={6}>
